Renamed vc4gfx to vcgfx
