+++ arch/arm-native/soc/broadcom/2708/hidd/vc4gallium/vc4_galliumclass.c
@@ -36,14 +36,14 @@
 #include "vc4_drm_aros.h"
 #include "vc4_v3d.h"  /* provides ARM_PERIIOBASE, bcm2708.h, GPU_BUS_ADDR */
 
-/* vc4gfx bitmap attributes — must match the aoHidd_VideoCoreGfxBitMap_*
- * enum in vc4gfx_bitmap.h */
+/* vcgfx bitmap attributes — must match the aoHidd_VideoCoreGfxBitMap_*
+ * enum in vcgfx_bitmap.h */
 #define aoHidd_VideoCoreGfxBitMap_Drawable      0
 #define aoHidd_VideoCoreGfxBitMap_BackDrawable  1
 #define aoHidd_VideoCoreGfxBitMap_Flip          2
 #define aoHidd_VideoCoreGfxBitMap_Overlay       3
 
-/* Mirrored from vc4gfx_bitmap.h, like the attr indices above. */
+/* Mirrored from vcgfx_bitmap.h, like the attr indices above. */
 struct vc4gfx_overlay
 {
     ULONG ovl_Phys;
@@ -562,11 +562,11 @@ void vc4_aros_wait_idle(struct vc4galliumstaticdata *sd)
 }
 
 /*
- * Scanout page query for zero-copy fullscreen GL. Resolves the vc4gfx
+ * Scanout page query for zero-copy fullscreen GL. Resolves the vcgfx
  * framebuffer's two flip pages and announces them (via sd->scanout_*)
  * as acceptable GEM_OPEN names, so the Mesa side can wrap them as BOs
  * and render straight into the back page. Returns 0 on success, -1
- * when the bitmap isn't a flippable vc4gfx framebuffer.
+ * when the bitmap isn't a flippable vcgfx framebuffer.
  */
 int vc4_aros_get_scanout(struct vc4galliumstaticdata *sd, OOP_Object *bm_obj,
                          struct vc4_aros_scanout *out)
@@ -606,7 +606,7 @@ int vc4_aros_get_scanout(struct vc4galliumstaticdata *sd, OOP_Object *bm_obj,
 }
 
 /*
- * Flip the vc4gfx framebuffer pages after the GPU has rendered a full
+ * Flip the vcgfx framebuffer pages after the GPU has rendered a full
  * frame into the back page. Waits for the GPU (and any in-flight blit
  * DMA) before the page becomes visible. Returns 0 on success.
  */
@@ -775,7 +775,7 @@ void vc4_aros_clear_overlay(struct vc4galliumstaticdata *sd,
 /*
  * Blit a rectangle from a BO to an AROS bitmap. Waits for GPU work
  * to finish first; pins/unpins the BO around the access; tries DMA
- * to the framebuffer phys address when the dest is a vc4gfx bitmap,
+ * to the framebuffer phys address when the dest is a vcgfx bitmap,
  * falls back to WritePixelArray otherwise.
  *
  * Caller supplies resolved fields — no Mesa types cross the API.
@@ -868,7 +868,7 @@ void vc4_aros_display_blit(struct vc4galliumstaticdata *sd,
 
     _t_after_v3d = VC4G_NOW_US();
 
-    /* Fast path: DMA into a vc4gfx framebuffer phys address. */
+    /* Fast path: DMA into a vcgfx framebuffer phys address. */
     {
         IPTR fb_addr = 0;
         IPTR fb_bpr = 0;
@@ -983,7 +983,7 @@ void vc4_aros_display_blit(struct vc4galliumstaticdata *sd,
                     width, height, dst_x, dst_y, bm_obj);
             }
 
-            /* Fallback for non-vc4gfx bitmaps — CPU reads BO data,
+            /* Fallback for non-vcgfx bitmaps — CPU reads BO data,
              * so we must flush ARM D-cache to see fresh GPU output. */
             CacheClearE(src_addr, src_stride * height, CACRF_ClearD);
