+++ arch/arm-native/soc/broadcom/2708/hidd/vcgfx/vcgfx.conf
@@ -10,7 +10,7 @@ options         noexpunge
 ##end config
 
 ##begin cdefprivate
-#include "vc4gfx_hidd.h"
+#include "vcgfx_hidd.h"
 ##end cdefprivate
 
 ##begin methodlist
