+++ arch/arm-native/soc/broadcom/2708/hidd/vcgfx/vcgfx_bitmap.h
