+++ arch/arm-native/soc/broadcom/2708/hidd/vcgfx/vcgfx_sdtv.c
@@ -5,7 +5,7 @@
 #define DEBUG 0
 #include <aros/debug.h>
 
-#include "vc4gfx_hidd.h"
+#include "vcgfx_hidd.h"
 
 int FNAME_SUPPORT(SDTV_SyncGen)(struct List *modelist, OOP_Class *cl)
 {
