+++ arch/arm-raspi/boot/mmakefile.src
@@ -83,7 +83,7 @@ ARM_BSP := aros-$(AROS_TARGET_CPU)-bsp.rom
 #MM kernel-bwfm-bcm2708-quick \
 #MM hidd-i2c-quick \
 #MM kernel-hidd-gfx-quick \
-#MM hidd-vc4gfx-quick \
+#MM hidd-vcgfx-quick \
 #MM kernel-sdcard-quick \
 #MM kernel-hidd-input-quick \
 #MM kernel-hidd-kbd-quick \
@@ -140,7 +140,7 @@ ARM_BSP := aros-$(AROS_TARGET_CPU)-bsp.rom
 #MM kernel-sdcard \
 #MM hidd-i2c \
 #MM hidd-i2c-bcm2708 \
-#MM hidd-vc4gfx \
+#MM hidd-vcgfx \
 #MM kernel-hidd-gfx \
 #MM kernel-hidd-input \
 #MM kernel-hidd-kbd \
@@ -179,7 +179,7 @@ PKG_RSRC_ARCH := processor battclock
 PKG_DEVS      := input gameport keyboard console sdcard USBHardware/usb2otg
 PKG_DEVS_ARCH := timer
 PKG_HANDLERS  := con ram cdrom sfs fat afs
-PKG_HIDDS     := hiddclass inputclass gfx keyboard mouse i2c i2c-bcm2708 vc4gfx
+PKG_HIDDS     := hiddclass inputclass gfx keyboard mouse i2c i2c-bcm2708 vcgfx
 PKG_CLASSES   := USB/hub USB/massstorage
 
 %make_package mmake=kernel-package-raspi-arm file=$(AROSDIR)/$(ARM_BSP) \
@@ -241,12 +241,14 @@ $(TARGETDIR)/core.bin.o: $(OSGENDIR)/boot/core.elf
 #MM
 distfiles-raspi-armbe : $(AROSDIR)/aros-armeb-raspi.img  $(AROSDIR)/$(ARM_BSP) $(AROSDIR)/config.txt
 	@$(ECHO) Copying RasPi-BE distfiles...
-	@$(CP) -R $(AROSDIR) $(DISTDIR)/
+# Copy the contents, not the directory - see the aarch64 rule.
+	@$(CP) -R $(AROSDIR)/. $(DISTDIR)/
 
 #MM
 distfiles-raspi-armle : $(AROSDIR)/aros-arm-raspi.img  $(AROSDIR)/$(ARM_BSP) $(AROSDIR)/config.txt
 	@$(ECHO) Copying RasPi distfiles...
-	@$(CP) -R $(AROSDIR) $(DISTDIR)/
+# Copy the contents, not the directory - see the aarch64 rule.
+	@$(CP) -R $(AROSDIR)/. $(DISTDIR)/
 
 #MM
 clean ::
